[fidl][libfuzzer] Remove HLCPP based fuzzer
It hasn't been used for a long time and is still based on the deprecated
HLCPP bindings. It's superseded by the realm fuzzer RFC-0117. If we want
to bring it back based on new bindings we can refer to this CL.

Bug: 408267642
Bug: 42070033
Bug: 42075797
Bug: 42151687
Bug: 42052603
Bug: 42052569
Change-Id: Ic0605e7becf70eef83f5ac78b196f0528121a9ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1244405
Reviewed-by: David Dorwin <ddorwin@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Alexandre Zani <azani@google.com>