[fidl][powerlevel] Introduce fuchsia.hardware.powerlevel
Typically the power level for a particular piece of hardware is
configured by requesting a specific voltage to be supplied to that piece
of hardware. In those cases fuchsia.hardware.vreg is a good choice of
FIDL interface.

Some hardware only provides an interface that permits the setting of an
opaque power level. The specific voltages, currents, clock frequencies
that this power level corresponds to is a property of the
implementation.

This FIDL interface provides a mechanism to set the power level for
hardware that only provides an opaque power level as its interface.

Fixed: b/407095566
Test: Linked against a driver and built
Change-Id: Id338b08d31cc8cbc6d6b629d33d61ec6367cb23d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238719
Reviewed-by: Andres Oportus <andresoportus@google.com>
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>