[kernel][vm] Refactor loaned page replacement
Moves the loaned page replacement in ProtectRangeFromReclamationLocked
to be outside the main lock acquisition. This is necessary since the
loaned page is being replaced on a potentially different VMO to what is
otherwise being operated on and so will need a different lock
acquisition and DeferredOps.

Bug: 338300943
Change-Id: I67c22cb21198e4da0d8d0dbea4471330017f1fd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242585
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>