[kernel] Add Drain method to SpscBuffer
KTrace allows users to "Rewind" the trace buffer, which effectively
means emptying it. This CL adds support for this functionality to
the SpscBuffer with a method called Drain, which is logically
equivalent to performing a Read into /dev/null. Callers could have
already implemented this function themselves by calling Read in a
loop with a no-op CopyOutFn; adding this to the API just makes the
operation more efficient.

Bug: 404539312
Change-Id: I3ec8bf9a18c0ad870fa48e7dec192e44ca8b31ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237565
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>