[ffx][log] Adding symbolizer as a data_dep
This adds //tools/symbolizer as a data_deps for the `ffx log` subtool.
The subtool invokes `symbolizer` and if it is not built and installed as
a host_tool, ffx cannot find it in-tree.

Out of tree, the sdk manifest is used to locate it.

Bug: 409133313
Change-Id: Ia1bd449e92ab015a6e9be4cfcd98e4da14a1392e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1248005
Fuchsia-Auto-Submit: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>