[build][python] Fix typing issues in assembly modules.
This removes typing issues in the module files, but also
in some of the scripts that use them.

Note that the @serialize_json decorator is *not* compatible with
mypy (as it breaks static analysis), so is no longer used in
the definitions of ImageAssemblyConfig and AssemblyInputBundle.
A comment has been added to warn about it usage in serialization.py.

+ Remove broken commands in assembly_input_bundle_tool.py
  which relied on non-existent methods in AssemblyInputBundle
  and thus would fail at runtime ("diff", "intersect" and
  "find_blob").

Bug: 366268742
Change-Id: I7adb8b6e551bd35ab8331a82711a86b827eee697
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241526
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: David Turner <digit@google.com>