[display][coordinator] Migrate integration test to Virtcon fallback.
This CL updates IntegrationTest.SendVsyncsAfterImagelessConfig to wait
for the virtcon client to gain display ownership after it connects to
the Display Coordinator. This ensures the client has ownership before it
calls ApplyConfig(), which is the path we intend to test.

Bug: 388885807
Test: fx test //src/graphics/display/drivers/coordinator
Change-Id: I422d5d0f1846a8d47eaf3f91babcc198b3c69aca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1244566
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>