[kernel] Change ktrace categories to use acq/rel semantics
It turns out that my reasoning for making the load and store of the
categories bitmask relaxed in https://fxrev.dev/1243384 was wrong.
Specifically, the order in which we set the categories bitmask is
important, because we want arbitrary trace records to enter the
ktrace buffer only after we've finished emitting the thread and
process metadata.

Bug: 404539312
Change-Id: Ib31947cbc92bc08d6444a376edad2719dc61d59f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1248026
Reviewed-by: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>