[vfs] Remove dependencies on the rust vfs
The rust vfs was converted to a dylib which caused performance
regressions in fxfs. Ideally, fxfs should only depend on the vfs
statically. Fxfs depends on severals crates which depend on
fuchsia_component which depends on the vfs dylib. fuchsia_component
only depends on the vfs for ServiceFs which none of the crates that fxfs
depends on make use of. This change switches fxfs' dependencies to only
depend on the client portion of fuchsia_component
(fuchsia_component_client).

Bug: 396172331
Change-Id: Iba8f86b16467138d40ce33e198c0388b5990fefb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235845
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>