[archivist] Add e2e benchmark
These two benchmarks validate the time it takes from emitting a log
until we read it and parse it back from archivist. It's possible we may
want to remove the parsing side of it, but leaving it there for now,
although it may inflate the numbers a bit.

Bug: 300156427
Change-Id: I2ab75907ee52a4b50c3ddac2c1a43312f575f4ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1247304
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>