[wlandevicemonitor] Lock inside destroy_iface.
Prior to this change, it was possible for a future
communicating with a PHY driver to destroy an interface
and a future monitoring the interface for liveness to race
to complete.  When the monitoring future won the race, the
rest of the system was notified that an interface had been
removed.  If another component attempted to create an
interface at this point, the operation would fail because
the destruction process was still ongoing in the PHY driver.

This change introduces an async mutexing scheme around
calls to destroy_iface.  In the case where an interface is
being destroyed by an API request, listeners will not be
notified of interface removal until the driver has
responded to the destruction request.

Bug: 370381717
Test: fx test wlandevicemonitor-tests
Test: fx test hw-sim-tests
Multiply: wlandevicemonitor-tests
Multiply: wlan_policy_ap_test_fc: 500
Multiply: wlan_policy_test_fc: 500
Multiply: wlan_test_fc: 500
Change-Id: I4b21eedc2edff91ae5707bc627dc1122dbf624a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238716
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>