[sherlock] remove gpt type guid remapping
This metadata block would tell the gpt driver to remap partitions with
certain labels to have specified type guids. The two consumers of
partition metadata are fshost and the paver. Fshost doesn't use type
guids for identifying partitions.

The paver does, mostly via FindPartition, but there is an issue with the
implementation - FindPartition reads the gpt metadata off the disk so it
ignores the remapping (which is at the driver level), but then it passes
the non-remapped type guid into OpenBlockPartition, which _does_ use the
remapped type guid to find the partition and open a connection.
Therefore, if the remapping applied to any partitions in use, OTAs would
already be broken for sherlock. Since they aren't, we can assume the
remapping is a no-op.

See the bug for a slightly more detailed discussion.

Bug: 393194863
Change-Id: Ief49022f267b13f7ffa7d875bba1d07925ee831a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238668
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>