[pkg] Only use BlobReader/BlobWriter to access blobs.
Removes the conditional use in pkg-cache main for choosing between using
the old fuchsia.io protocols and the new reader/writer protocols.

Cleans up all the production uses of blob opening with fuchsia.io, and
updates all the integration tests that relied on pkg-cache behaving this
way.

Does not clean up all the optionals and branches for the number of
hermetic tests that explicitly call the fuchsia.io paths or create mocks
assuming that the fuchsia.io paths are available.

Bug: 408439508

Change-Id: Ifc3dc87811dff5a78c97ae76f9f81b39c8b8aee5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243164
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>