[starnix] Consume all ComponentController requests
We were only consuming the first ComponentController request in the
stream, which means that if a Stop request came in, and then timed out,
and subsequently a Kill request came in, we'd never observe the latter.

Address this by consuming the requests in a loop so that even if we've
already started shutdown due to a Stop request, we still respect
subsequent Kill requests.

Bug: 409364939
Test: _session_restart_test
Change-Id: I27adc8a4a7f4fcd4d92bb75a2845bfce1c17cefd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249005
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>