[kernel][vm] AddNewPagesLocked directly cleans up
Changes AddNewPagesLocked to directly removes pages from the page list
when cleaning up in the failure path instead of using
DecommitRangeLocked. The generality of DecommitRangeLocked is not
needed here, and in AddNewPagesLocked we are unable to, and know that we
do not need to, perform a range change update over our children, which
DecommitRangeLocked wants to perform.

Bug: 338300943
Change-Id: Ifb1ad4917f2bd36f7c3414e0e5079ec3d466900f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242590
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>