[starnix] Send attribution update on process name changes
With this change, Starnix will send a memory attribution update when the
name of one of its Linux processes changes.

This change also refactors some of the memory attribution code. In
particular, the memory attribution module is event-driven only, and no
longer wakes up every 5 seconds to scan the list of processes.

This is a reland of fxr/1217147. fxr/1217147 held the PidTable lock for
the whole duration of the event processing loop, including the sleep,
which was incorrect and degraded performance.

BUG=393078902

Change-Id: I21e562bf6e4df9f9fa9c617e93d81d041a487537
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239525
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Étienne J. Membrives <etiennej@google.com>