[persistence] Use B-Tree to group tags by time
Refactor Scheduler to use a B-Tree while grouping tags that need to be
fetched in the future. This is meant to be a readability improvement.
While this also improves over the existing O(N^2) solution with
O(N log N), it may not result in better performance for small values of
N due to the constant time for creating a B-Tree.

Change-Id: I2e74f2ad30cadf1863f1cde3f195f4f2155cc1ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1247458
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Sam Balana <sbalana@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>