[dma-buffer] enable uncached contiguous buffers
Zircon has supported multi-page contiguous uncached VMOs for several
years, but the dma-buffers BufferFactory did not provide access to the
functionality.

This change adds a bool, much like on BufferFactory::CreatePaged, to
enable or disable caching on new ContiguousBuffers. As this is a
signature change, all callers are updated to call with caching enabled
-- the prior default.

Bug: 409561709
Change-Id: Ieed73c987d2a4ad86d7f8be3baa37ece62df451b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249004
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Will Drewry <drewry@google.com>