[rust][bt] Add bt-gatt and bt-common
Add mirror of bluetooth.googlesource.com, which is a repository for
developing Bluetooth crates, as well as a script to update the mirror
to a specific git reference.

This mirroring is temporary and meant to allow changes in
fuchsia.googlesource.com to happen simultaneously with a mirror
reference change. LICENSE file symlinks are copied in the mirror.

Update the bluetooth.googlesource.com mirror to commit
c88c57d26787871904541c1beb7881f6e4b18b66

Add the bt-common and bt-gatt crates to the rust_crates packages
in prep for later CLs.

Bug: 42086536
Test: build changes only, fx build w/new packages
Change-Id: I0d906d0a87097b8e0fdeba181250ffba6468c75b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001973
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>