[uart] Introduce uart::all::Config::Match
Introduce a static method in charge of returning a matching
configuration object if any. This allows to slowly migrate more of the
logic that uses uart::all::Driver as a payload to be moved/copied
around.

Removes `MaybeCreate` from drivers in favor of `TryMatch` which returns
a configuration object. uart::all::Driver::Match will be eventually be
retired and it just relies on uart::all::Config::Match to determine
which driver to use.

Bug: 409114044
Change-Id: Id0ca4c3cbc2dca06824cd7ee6fb8f4bf6d45ba2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1247449
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>