[bazel] Fix bazel_action.py query cache key computation.
Use the content of input starlark files used for cqueries to compute
the key used for the query cache. Without this, modifying these files
would end up with stale cache entries being used in the next incremental
build.

This is normally not a problem for most incremental builds, because
any change to build files would re-run `fx gen` which wipes the
cache entirely, but a change that only changes the starlark files
would trigger this bug, so fix it there, as it was discovered
during local development.

Change-Id: I94e3812607e941fadf483a4633e9a1ac9dfc65a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249944
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>