[ffx][flash] fix flashing sparse images
Now detect if an image in the product bundle
is already a sparse image and is greater than the
max download size of the Target. If that is the case,
expand the image to a temp file and then re-sparse
the image to a new set of sparse images, then flash
the re-sparsed images.

Bug: b/390670098
Bug: b/394699849
Fixed: b/390670098
Fixed: b/394699849

Change-Id: Ifee51170d948dee350354ba72825c0c159c39a6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243987
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>