[starnix] Fix race in thread and task release
When a Task was released, its WeakRef could not be updated anymore,
while the TheadGroup that was owning it was not updated yet and could
still be accessed through the pid table. This ended up creating
ThreadGroup with no available task in.
This change the Task release to require a lock to the pid table. It
ensures that while holding the lock to the pid table, any thread group
read from it will have at least one active task.

Fixed: b/409478205, b/409425986

Change-Id: Iae8660f3501adf82037132a1c296110eb73855d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1247587
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>