[fxfs] Cache errors when replaying profiles and downgrade log level
This gets noisy for something that is actually quite acceptable. It was
originally WARN level on the assumption that after an OTA none of the
files should actually have gone missing yet, but that can be racy with
whatever the SWD stack might be up to, so it might not indicate an
actual problem.

Adds an info log to indicate count of success and failure, so if this
becomes suspiciously high we'll have an indicator without creating a lot
of log spam.

Change-Id: I9169420b94078af0deb3bfa084deff072f92e2b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249144
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>