[ffx] Populate ssh host address even without an ssh host pipe
We need to know the address of the host from the target's perspective to
make package management work. Usually we get this as a side effect of
establishing an SSH host pipe, but with USB or VSOCK we don't actually
do that, leading to a surprising breakage.

Now we have a separate path where we can SSH in *just* to get the host
address without starting a pipe. We should only hit this path when we
don't have an SSH host pipe so existing workflows shouldn't change.

Fixed: 1250227
Change-Id: I4c1352a4d117e401ac0adc63ce29fa52fe6bbd17
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250227
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>