[kernel][vm] PageLookup uses LockedPtr
PageLookup uses a LockedPtr for the owner instead of a raw VmCowPages*.
FindPageContentLocked takes the required ancestor lock and populates
LockedPtr if an ancestor owns the page. Otherwise the LockedPtr is left
unset and the caller can assume that |this| is the page owner, which it
already holds locked.

LookupCursor functions change accordingly to use this new LockedPtr. One
consequence is that the owner() helper is no longer valid and is
removed.

Bug: 338300943
Run-All-Tests: true
Change-Id: I346b83a75c568d481a10d6603f5c5cbbff3b0300
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1249114
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>