[kernel][vm] Refactor PageFaultLocked
Restructures the PageFaultLocked code to first cast to the underlying
VMO type, and then take a lock on that object. This allows for the
VmObjectPaged case to create a DeferredOps outside the lock, but only
if it's a paged object. This requires some small duplication of logic
inside each of the if/else branches.

Bug: 338300943
Change-Id: If5acbc9df56083b68201c68dcb988f864a680a31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242596
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>