[audio][test] Make drivers tests 25% faster
In AudioDriverTests, at the end of each FIDL exchange with a
driver (i.e. when the test case is complete), we always wait
for additional FIDL activity such as an unexpected callback,
or a disconnect. Rather than just "waiting for something bad
to NOT happen", we can be proactive and use a simple "call &
response" to ensure that the driver is still responsive over
FIDL - we use GetHealthState for this purpose.

By moving to this, eliminating our 100ms wait per test case,
this shortens the execution time for AudioDriverTests suites
to 75% of the previous run-time.  \o/  yay!

Bug: 370741565

Change-Id: Iebcf0f5fa848238f10ebc2f789311ee5b7c29465
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1251224
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>