[audio] ADR eager-start config (1/3)
Add a platform configuration flag to specify whether the
AudioDeviceRegistry service should be eagerly started.

A flag is being added to the platform configuration schema
for audio. Because out-of-tree product definitions include
settings for audio, a soft transition is required. This CL
is the first of three that are ultimately needed.

Bug: 390573041
Bug: 410822855

Change-Id: I59f7ba20ce60d68fdf10bcf448db75e7ed733a97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228525
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>