[blobfs] Check if the preexisting BlobWriter is still active.
Verifies that a preexisting BlobWriter is actually still active by
calling `signal_peer()` on the channel to see if the client side has
closed but the executor just hasn't handled the message yet. This
provides protection against racing with the old client handle close
and the new BlobWriter creation.

Adds bare pointers as back-refs from the associated blobs to their
BlobWriters. This lets duplicate blobs found in the cache to allow
manipulation of the BlobWriter. In particular, it allows dropping the
Blob so that it gets recycled out of the cache and makes room for the
new version being created.

Bug: 408439508
Change-Id: Ia231c45e18224e135297c6c08698cf4faf6fbddf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250905
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Suter <csuter@google.com>