[symbolizer] Change log level to debug for symbolizer
I made this change because when `ffx profiler symbolize` be called, a
lot of non-fatal errors been print out to the screen for the profiler
users. A lot of backtraces be handled at once, and of course, some of
them might fail, and users are not care about the detailed failure
reason.

Bug: 409569703

Change-Id: I1cc1e5297e370412dac0ce4e2408f51007bf7db6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1251015
Commit-Queue: Jia Li <lizhijia@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>