[bazel] Fix `fx bazel-tool set_gn_targets <label>`
This tool command was broken by a previous CL [1] that added new
fields in the bazel_build_actions metadata of bazel_action() GN
targets, which broke gn_target_utils.py, invoked at runtime.

This was undetected because this is a development tool that is
only used by the build team, and never runs at build or test
time, ether locally or in infra.

This fixes the situation by:

- Adding dataclass field definitions for the extra metadata keys
  in gn_target_utils.py

- Documenting the new keys in //BUILD.gn

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1220864

Change-Id: Id7eb1873ba07bf40486f4b0bf2bd4f6633f04e27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253186
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Fang <fangism@google.com>