[kernel][arm64] Change arm64 spinlocks to be 32bit
This matches the size of the other two architectures, which use
cpu_num_t (uint32_t) as the basic spinlock type.

Was previously using unsigned long for historical reasons, to make it
match the size of the architecture when building with arm32 and arm64
at the same time.

No functional change except for the size of the spinlock and any
alignment or size changes this may entail.

Change-Id: I7f1ca24f747ae77d67229a880a0d421fd48669bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1251013
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>