[ffx][discovery] fix dangling mDNS query loop
As opposed to `Task::local(...).detach();`, now keep track of the
recv tasks for IpV4 and Ipv6 loops.

Note: `detach()` will cause the thread to not be stopped when the task
is dropped.

This will stop the Tasks (and therefore the dangling mdns traffic) that
the mdns_discovery lib was creating.


Change-Id: Idbe09eeea51c787c44f7499ae09ae0cb6a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253433
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>