[thread-sampler] Add asserts for per cpu structs
Timer migration could be a possible cause of the panic we see in the
thread sampler code. We don't expect timer migration to run in a normal
profiling scenario, but as the sampler code is currently incompatible
with timer migration, we should at least assert some invariants to
ensure it doesn't happen.

Bug: 410034543
Bug: 392888672

Change-Id: I4c8e01904c0a8a043636916698a30bf184eeaf85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250734
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>