[bt] Add one-to-one matcher for hanging get.
This change adds a library struct to match pairs of things that can
become available in any order.  Clients can enqueue to the left or the
right queue, and when there is an item available in each queue, a stream
will yield a result.

Tests: fx test one-to-one-tests

Change-Id: I8f120a2f78282ab1b7d16399535de092ea846a88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926955
Commit-Queue: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>