[trace][starnix] Fixing ATrace End parsing
According to the Perfetto systrace parser code, it is allowable to
have an End event for a thread scoped event that does not contain the
TGID. The reasoning behind this is since it is a thread scoped, the id
is not needed to match up with the Begin event.

We currently use the pid to manage the event queue, so in the case of
End missing the TGID, the current task pid is read to find the correct
queue.

Bug: 401645216
Change-Id: I24b0f182af095925e3cf52d734dc7362b0f66eb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253428
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>