[graphics][drivers] Add a TestService
Update the test drivers to expose a unique service that gives
access to the test protocols. This means that we can be honest
in the CML and FIDL type system about what capabilities the
drivers are exposing, and we can remove the templated FIDL
protocol serving from the drivers.

This is a soft transition as OOT drivers will need to be updated
before the old functions can be fully deleted.

Bug: b/410611164

Change-Id: I22a49008b39063fe32046b0616fcd5487e6565fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1252545
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>