[ffx][flash] Improve flashing speed
We used to use a buffer which is 4KB, this is proven to be a bottleneck
of flashing performance. Now we change it to read file into a 50MB
buffer, and let the TCPStream to handle the chunking. From local
testing, the chunk sizes is about 1.6MB on average. From local testing
with pontis, this decrease the flashing time of large image from 279
seconds to 75 seconds.

Also we noticed the set_active call can fail on sorrel, we change it to
be best effort and log a warning when fail. Also added a 3 seconds
timeout on set_active call, as this call is not computationally
expensive and should response pretty fast when succeed.

Bug: 409544982
Change-Id: I138467b62f00c8193684c1109969afa4e1b789be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1252985
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jiaming Li <lijiaming@google.com>