[wlan][policy] Use global backoff, stationary roam
This change updates the roam scan backoff logic in the stationary roam
profile to use a simpler global backoff. This is because, at release,
there will only be a single roam reason, so much of the existing backoff
(e.g. extended only when there are not new roam reasons) is irrelevant.

Eventually, we may need a backoff solution that tracks multiple reasons
again, but there are some edge cases to consider. That effort is tracked
in the bug below.

Bug: b/409839414
Test: wlancfg tests update to reflect new backoffs

Change-Id: I02d967d9ee33b9dd77a7665b7babb16c033fcd30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1254547
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>