[starnix] Mark sockets when feature enabled
Before this change, starnix has a local state for SO_MARK which does
not affect routing. In this change we want SO_MARK to be used by the
netstack, but put it behind the feature flag so that there are no
user visible behavior changes in routing.

The feature is enabled for the debian container so that syscall tests
can test against the change.

Bug: 410576727
Change-Id: I9eb9c0b9c4fd04ed34303ebcd785c39f8ad96529
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238848
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Sergey Ulanov <sergeyu@google.com>