[update-rustc-third-party] Download cargo-bazel
... instead of building it with crate_universe bootstrap. The bootstrap
process creates a new bootstrap repository and does a full rebuild of
cargo-bazel every time, which can take around a minute of even longer.
This change switches to download this binary from upstream release with
a matching version. As of version 0.59.1 today, the binary is < 20MB,
so should be pretty fast to download.

Note Mac support is included in this change but not tested.

Change-Id: Id33e44105fd5e35e0f9297fb9e2786b50bfc0a7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1253005
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Jay Zhuang <jayzhuang@google.com>